còn thiếu link và thích ứng màn hình

+++ data/admin/remove_group.php
@@ -2,14 +2,21 @@
     require "../connect/connect.php";
 
     $id = $_POST["id"];
+    $file = $_POST["img"];
 
-    $sql = "delete from groups where id='$id'";
-    $kt = mysqli_query($conn, $sql);
-    if($kt){
-        echo 1;
-    }
-    else {
+    $pathName_sl_remove= "../../public/img/leader/".$file;
+    if(!unlink($pathName_sl_remove)){
         echo 0;
+    }
+    else{
+        $sql = "delete from groups where id='$id'";
+        $kt = mysqli_query($conn, $sql);
+        if($kt){
+            echo 1;
+        }
+        else {
+            echo 0;
+        }
     }
 
     mysqli_close($conn);
